refactor: 즐겨찾기 삭제 메서드 Controller 변경
